pyproject.toml
[tool.poetry]
name = "pyproject-toml"
version = "0.1.0"
description = ""
authors = ["Nurlan <nurlan.grachyov@mail.ru>"]
readme = "README.md"
packages = [{include = "pyproject"}]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
openpyxl = "^3.1.5"
python-dotenv = "^1.0.1"
pandas = "^2.2.3"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.11.2"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^5.0.0"

[tool.mypy]
disallow_untyped_defs = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
line_length = 119

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.report]
exclude_also = [
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]
